[Upd] Moved file to independent app
